Starting point for new Pool implementation
This replaces the guts of the old Pool implementation with one that
takes no locks, and thus should never block the reactor (#64). Instead,
a "recycler" future is spawned that receives any discarded connections,
do whatever cleanup is necessary on them, and then send them back to the
clients over a lock-free mpsc-like channel. There's also a fast-path for
returning pristine connections that are dropped to the pool directly.