opts: Fix IPv6 handling in MySQL URL parsing
MySQL URL parsing in `from_url_basic()` handles IPv6 incorrectly so that
if, for example, an URL like `mysql://[::1]/dbname` is passed, the
`Opts.inner.ip_or_hostname` will end up with the value `"[::1]"` since
the the `url.host_str()` simply returns the whole host part as is,
including brackets in this case.

But later this host is used in `conn/mod.rs` as:
  Stream::connect_tcp((opts.get_ip_or_hostname(), opts.get_tcp_port()))

and `connect_tcp()` doesn't expect IPv6 to contain brackets since this
is standalone param. As the result it treats `"[::1]"` as a domain, not
IPv6, tries to resolve it and expectedly fails.

Fix it by switching from `url.host_str()` to `url.socket_addrs()` that
works with IPv6 rather than with strings.

`url.host()` unfortunately can't be used here since it has similar
problem by working with string for domain case w/o a try to resolve it.

The change is covered by a bunch of new tests and verified by:
  cargo test opts::test

The tests can be used to reproduce the problem.

Signed-off-by: Andrey Ignatov <rdna@fb.com>