Completely shut down connections on disconnect
Previously, connections that got replaced (due to `reset` or an upgrade
to `Socket`) were simply dropped and left to be cleaned up by spawning
the disconnect routine. This patch makes them be torn down when the new
connection is established instead.

This patch also ensures that we completely close the underlying stream
during cleanup, to minimize the time during which we hold up resources
as the server.