Accept either bytes or strings for queries
As discussed in #194, MySQL actually accepts arbitrary sequences of
bytes, not just utf-8 strings, as queries, but this crate is limited to
only working with types that impl AsRef<str>. To allow sending arbitrary
byte slices as queries this commit, which goes along with and requires
https://github.com/blackbeam/rust_mysql_common/pull/64, introduces a new
`AsQuery` trait, which is impl'd for all of the standard library types
that either impl AsRef<str> or AsRef<[u8]>, and uses that trait in place
of `AsRef<str>` for all query methods, going on down the chain to error
types, internal cache structures, etc. as well.

Fixes: #194